requirements.txt
streamlit>=1.32.0
openai>=1.24.0,<2.0.0
langchain==0.1.12
langchain-openai==0.1.7
langchain-community>=0.0.29
faiss-cpu>=1.7.4
tiktoken>=0.6.0
pypdf>=4.1.0
python-dotenv>=1.0.1
plotly>=5.20.0
pandas>=2.2.1
pydantic>=2.5.0
pytest>=7.0

wordcloud>=1.9.3
spacy>=3.7.0,<4.0.0
networkx>=3.2.0
pyvis>=0.3.2
matplotlib>=3.8.0

textstat>=0.7.3
nltk>=3.8.1
language-tool-python>=2.7.1
textblob>=0.18.0
numpy>=1.26.0,<3.0.0
